+++ src/Spreads.Collections/MapCursor.fs
@@ -160,7 +160,4 @@ type MapCursor<'K,'V when 'K : comparison>
     member this.Source with get() = this.Source
     member this.Clone() = this.Clone()
     member this.IsContinuous with get() = this.IsContinuous
-    member this.TryGetValue(key, [<Out>]value: byref<'V>) : bool = 
-      let entered = enterLockIf syncRoot isSynchronized
-        try
-        map.TryGetValue(key, &value)
+    member this.TryGetValue(key, [<Out>]value: byref<'V>) : bool = map.TryGetValue(key, &value)